refactor: fix routes names
